+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Fixed - for any bug fixes.
 Security - in case of vulnerabilities.
 
 ## [Unreleased] - XXXX-XX-XX
+### Fixed
+- Resource versions order
 
 ## [1.0.9] - 2018-12-03
 ### Fixed